Finished...Theoretically
So this code is right, but there's a bug in the Explosion.php file in PocketMine itself. It's been fixed and is in a pull request thanks to PEMapModder, and I have yet to test it with the fix. I'm currently waiting for shoghi to merge the fix and release a new dev build. Once I test, I'll update it in the PM Plugin Repo! :D